[gym/math] Promote generic SE3 math utilities from toolbox to common module.
